Generated documentation
There is an error on almost every file that the page level documentation
block is missing.  Best practices indicate that files declaring a class
should use a class level documentation block with no page level block. I
am unsure how to force phpdoc to ignore this error.

Also, the properties in the autoloader still need to be documented.